Fix auto restart bug

+++ auto_restart.py
@@ -1,7 +1,8 @@
-import discord
-from flask import Flask, request
-from threading import Thread
+import os
 import time
+
+import discord
+
 from modules.utils import BotUtils
 
 bu = BotUtils()
@@ -12,13 +13,14 @@ class AutoRestart(discord.Client):
         print('Logged on as {0}!'.format(self.user))
         self.channel = self.get_channel(bu.config['restart_channel'])
         while True:
-            fin = open('push', 'r')
-            if fin.read() == '1':
-                await self.channel.send('!r')
-                fin.close()
-                open('push', 'w').write('0')
+            if os.path.exists('push'):
                 fin = open('push', 'r')
-            fin.close()
+                if fin.read() == '1':
+                    await self.channel.send('!r')
+                    fin.close()
+                    open('push', 'w').write('0')
+                    fin = open('push', 'r')
+                fin.close()
             time.sleep(1)
 
     async def on_message(self, message):
